retriever/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
managed = true
package = true

[project]
name = "retriever"
version = "0.1.0"
description = "A modern RAG ingestion pipeline from Nvidia"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Jeremy Dyer", email = "jdyer@nvidia.com" },
]
license = { text = "Apache-2.0" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
]
dependencies = [
  "ray[data]>=2.49.0",
  "pypdfium2==4.30.0",
  "pandas>=2.0",
  "tqdm>=4.66.0",
  "typer>=0.12.0",
  "lancedb",
  #"nv-ingest-api==26.1.2",
  "fastapi>=0.114.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "requests>=2.31.0",
  "pydantic>=2.8.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pillow>=12.0.0",
  "numpy>=1.26.0",
  "python-multipart>=0.0.9",
  "transformers>=4.49.0",
  "tokenizers>=0.20.3",
  # TensorRT support (keep Torch/TorchVision versions aligned; mismatch breaks torchvision C++ ops like nms)
  "torch~=2.9.0",
  "torchvision>=0.24,<0.25",
  "einops",
  "easydict",
  "addict",
  "nemotron-page-elements-v3==3.0.1",
  "nemotron-graphic-elements-v1==1.0.0",
  "nemotron-table-structure-v1==1.0.0",
]

[project.optional-dependencies]
dev = [
  "build>=1.2.2",
  "pytest>=8.0.2",
]

[project.scripts]
retriever = "retriever.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
